+++ trunk/MyGUIEngine/src/MyGUI_SimpleText.cpp
@@ -145,7 +145,7 @@ namespace MyGUI
 
 
 		// устанавливаем размер текста
-		mContextSize.set(width, (float)mLinesInfo.size() * mFontHeight);
+		mContextSize.set(int(width), mLinesInfo.size() * mFontHeight);
 		mContextRealSize.set(mContextSize.width * mManager->getPixScaleX() * 2.0f, mContextSize.height  * mManager->getPixScaleY() * 2.0f);
 	}
